[modify]增加註解

+++ PhoenixCI/FormUI/Prefix3/W30340.cs
@@ -2,7 +2,6 @@
 using System.Windows.Forms;
 using BaseGround;
 using BusinessObjects.Enums;
-using BaseGround.Report;
 using System.Threading;
 using BaseGround.Shared;
 using Common;
@@ -25,23 +24,11 @@ namespace PhoenixCI.FormUI.Prefix3
 
          this.Text = _ProgramID + "─" + _ProgramName;
       }
-      public override ResultStatus BeforeOpen()
-      {
-         base.BeforeOpen();
-
-         return ResultStatus.Success;
-      }
 
       protected override ResultStatus Open()
       {
          base.Open();
-         emMonth.Focus();
-         return ResultStatus.Success;
-      }
-
-      protected override ResultStatus AfterOpen()
-      {
-         base.AfterOpen();
+         //讀取交易日
          emMonth.Text = GlobalInfo.OCF_DATE.ToString("yyyy/MM");
          emMonth.Focus();
          return ResultStatus.Success;
@@ -54,6 +41,9 @@ namespace PhoenixCI.FormUI.Prefix3
          return ResultStatus.Success;
       }
 
+      /// <summary>
+      /// 轉檔前檢查日期格式
+      /// </summary>
       private bool StartExport()
       {
          if (!emMonth.IsDate(emMonth.Text + "/01", "日期輸入錯誤")) {
@@ -71,7 +61,10 @@ namespace PhoenixCI.FormUI.Prefix3
          return true;
       }
 
-      protected void EndExport()
+      /// <summary>
+      /// 轉檔後清除文字訊息
+      /// </summary>
+      private void EndExport()
       {
          stMsgTxt.Text = "";
          this.Cursor = Cursors.Arrow;
@@ -80,19 +73,16 @@ namespace PhoenixCI.FormUI.Prefix3
          stMsgTxt.Visible = false;
       }
 
-      protected void ShowMsg(string msg)
+      /// <summary>
+      /// show出訊息在label
+      /// </summary>
+      /// <param name="msg"></param>
+      private void ShowMsg(string msg)
       {
          stMsgTxt.Visible = true;
          stMsgTxt.Text = msg;
          this.Refresh();
          Thread.Sleep(5);
-      }
-
-      private string OutputShowMessage {
-         set {
-            if (value != MessageDisplay.MSG_OK)
-               MessageDisplay.Info(value);
-         }
       }
 
       protected override ResultStatus Export()
@@ -101,16 +91,26 @@ namespace PhoenixCI.FormUI.Prefix3
             return ResultStatus.Fail;
          }
          string lsFile = PbFunc.wf_copy_file(_ProgramID, "30340");
+         //OutputShowMessage只會儲存ok的狀態,如沒有任何一個ok代表全部function都沒有資料
+         MessageDisplay message = new MessageDisplay();
          try {
             b30340 = new B30340(lsFile, emMonth.Text);
 
             ShowMsg("30340－「三十天期商業本票利率」期貨契約價量資料 轉檔中...");
-            OutputShowMessage = b30340.Wf30341();
+            message.OutputShowMessage = b30340.Wf30341();
             ShowMsg("30343－「三十天期商業本票利率」期貨契約價量資料(買賣方比重) 轉檔中...");
-            OutputShowMessage = b30340.Wf30343();
+            message.OutputShowMessage = b30340.Wf30343();
+
+            //連續跳2次無資料刪除檔案
+            if (string.IsNullOrEmpty(message.OutputShowMessage)) {
+               if (File.Exists(lsFile))
+                  File.Delete(lsFile);
+               return ResultStatus.Fail;
+            }
          }
          catch (Exception ex) {
-            File.Delete(lsFile);
+            if (File.Exists(lsFile))
+               File.Delete(lsFile);
             WriteLog(ex);
             return ResultStatus.Fail;
          }
