+++ PhoenixCI/FormUI/Prefix3/W30370.cs
@@ -2,12 +2,10 @@
 using System.Windows.Forms;
 using BaseGround;
 using BusinessObjects.Enums;
-using BaseGround.Report;
 using System.Threading;
 using BaseGround.Shared;
 using Common;
 using PhoenixCI.BusinessLogic.Prefix3;
-using DataObjects.Dao.Together.SpecificDao;
 using System.IO;
 /// <summary>
 /// john,20190305,年度期間法人機構期貨交易量統計表 
@@ -26,23 +24,11 @@ namespace PhoenixCI.FormUI.Prefix3
 
          this.Text = _ProgramID + "─" + _ProgramName;
       }
-      public override ResultStatus BeforeOpen()
-      {
-         base.BeforeOpen();
-
-         return ResultStatus.Success;
-      }
 
       protected override ResultStatus Open()
       {
          base.Open();
-         emMonth.Focus();
-         return ResultStatus.Success;
-      }
-
-      protected override ResultStatus AfterOpen()
-      {
-         base.AfterOpen();
+         //讀取交易日
          emMonth.Text = GlobalInfo.OCF_DATE.ToString("yyyy/MM");
          emMonth.Focus();
          return ResultStatus.Success;
@@ -55,14 +41,9 @@ namespace PhoenixCI.FormUI.Prefix3
          return ResultStatus.Success;
       }
 
-      private string OutputShowMessage {
-         set {
-            if (value != MessageDisplay.MSG_OK && value != MessageDisplay.MSG_NO_DATA)
-               MessageDisplay.Info(value);
-         }
-      }
-
-
+      /// <summary>
+      /// 轉檔前檢查日期格式
+      /// </summary>
       private bool StartExport()
       {
          if (!emMonth.IsDate(emMonth.Text + "/01", "日期輸入錯誤")) {
@@ -80,6 +61,9 @@ namespace PhoenixCI.FormUI.Prefix3
          return true;
       }
 
+      /// <summary>
+      /// 轉檔後清除文字訊息
+      /// </summary>
       protected void EndExport()
       {
          stMsgtxt.Text = "轉檔完成!";
@@ -89,6 +73,10 @@ namespace PhoenixCI.FormUI.Prefix3
          stMsgtxt.Visible = false;
       }
 
+      /// <summary>
+      /// show出訊息在label
+      /// </summary>
+      /// <param name="msg"></param>
       protected void ShowMsg(string msg)
       {
          stMsgtxt.Text = msg;
@@ -102,17 +90,27 @@ namespace PhoenixCI.FormUI.Prefix3
             return ResultStatus.Fail;
          }
          string lsFile = PbFunc.wf_copy_file(_ProgramID, "30370");
+         //message.OutputShowMessage只會儲存ok的狀態,如沒有任何一個ok代表全部function都沒有資料
+         MessageDisplay message = new MessageDisplay();
          try {
             b30370 = new B30370(lsFile, emMonth.Text);
             //wf_30371()
             ShowMsg($"30371－年度期間法人機構期貨交易量統計表 轉檔中...");
-            OutputShowMessage = b30370.Wf30371();
+            message.OutputShowMessage = b30370.Wf30371();
             //wf_30375()
             ShowMsg($"30375－年度期間法人機構期貨交易量統計表(維持率) 轉檔中...");
-            OutputShowMessage = b30370.Wf30375();
+            message.OutputShowMessage = b30370.Wf30375();
+
+            //連續跳2次無資料刪除檔案
+            if (string.IsNullOrEmpty(message.OutputShowMessage)) {
+               if (File.Exists(lsFile))
+                  File.Delete(lsFile);
+               return ResultStatus.Fail;
+            }
          }
          catch (Exception ex) {
-            File.Delete(lsFile);
+            if (File.Exists(lsFile))
+               File.Delete(lsFile);
             WriteLog(ex);
             return ResultStatus.Fail;
          }
